agri_ml_backend/requirements.txt
Flask
tensorflow
pillow
numpy
scikit-learn
opencv-python
pytesseract
spacy
transformers
joblib
python-dotenv